Refactor: main 관련 코드들 features 내부로 이동 (#46)
