+++ src/features/main/components/DiaryBottomSheetBackdrop.tsx
@@ -1,0 +1,5 @@
+import { BottomSheetBackdrop, type BottomSheetBackdropProps } from "@gorhom/bottom-sheet";
+
+export default function DiaryBottomSheetBackdrop(props: BottomSheetBackdropProps) {
+  return <BottomSheetBackdrop {...props} disappearsOnIndex={-1} appearsOnIndex={1} />;
+}
