+++ src/features/main/types/registerTypes.ts
@@ -1,0 +1,3 @@
+import type { SuccessDTO } from "@/src/types/commonTypes";
+
+export type RegisterDTO = SuccessDTO & { result: null };
