+++ src/types/calendarTypes.ts
@@ -1,14 +1,5 @@
 import { SuccessDTO } from "@/src/types/commonTypes";
 
-export type Day = {
-  year: number;
-  month: number;
-  day: number;
-  hashTag1: string;
-  hashTag2: string;
-  imageS3: string;
-};
-
 export type Diary = {
   year: number;
   month: number;
@@ -18,27 +9,9 @@ export type Diary = {
   summaryContent: string;
   hashTag1: string;
   hashTag2: string;
-  summaryMood:
-    | "HAPPY"
-    | "SAD"
-    | "TIRED"
-    | "ANGRY"
-    | "RELAX"
-    | "HAPPY"
-    | "SAD"
-    | "TIRED"
-    | "ANGRY"
-    | "RELAX";
-};
-
-export type MonthCalendar = SuccessDTO & {
-  result: Day[];
+  summaryMood: "HAPPY" | "SAD" | "TIRED" | "ANGRY" | "RELAX" | "HAPPY" | "SAD" | "TIRED" | "ANGRY" | "RELAX";
 };
 
 export type DiaryResult = SuccessDTO & {
   result: Diary;
-};
-
-export type DiariesResult = SuccessDTO & {
-  result: Diary[];
 };
